@ThomasdenH 2020-02-01
I have removed the last commit, which was the only one containing breaking changes.